*** EFM32 branch *** 1. Upgrade EFM32 driver libraries (CMSIS and efm32lib): version 2.3.0 -> 2.3.2 2. Upgrade EFM32GG_DK3750 development kit driver library: version 1.2.1 -> 1.2.2 3. Upgrade EFM32_Gxxx_DK development kit driver library: version 1.7.2 -> 1.7.3 4. Add LCD driver (SSD2119) and DEMO 5. Change SPI write format 6. Modify SD card driver due to SPI write format changed 7. Modify building scripts
git-svn-id: https://rt-thread.googlecode.com/svn/trunk@1883 bbd45198-f89e-11dd-88c7-29a3b14d5316

+++ bsp/efm32/Libraries/efm32lib/inc/efm32_ebi.h
@@ -2,7 +2,7 @@
  * @file
  * @brief External Bus Iterface (EBI) peripheral API for EFM32
  * @author Energy Micro AS
- * @version 2.3.0
+ * @version 2.3.2
  *******************************************************************************
  * @section License
  * <b>(C) Copyright 2010 Energy Micro AS, http://www.energymicro.com</b>
@@ -68,15 +68,15 @@ extern "C" {
  *******************************   DEFINES   ***********************************
  ******************************************************************************/
 
-#define EBI_BANK0    (1 << 1) /**< EBI address bank 0 */
-#define EBI_BANK1    (1 << 2) /**< EBI address bank 1 */
-#define EBI_BANK2    (1 << 3) /**< EBI address bank 2 */
-#define EBI_BANK3    (1 << 4) /**< EBI address bank 3 */
+#define EBI_BANK0    (uint32_t)(1 << 1) /**< EBI address bank 0 */
+#define EBI_BANK1    (uint32_t)(1 << 2) /**< EBI address bank 1 */
+#define EBI_BANK2    (uint32_t)(1 << 3) /**< EBI address bank 2 */
+#define EBI_BANK3    (uint32_t)(1 << 4) /**< EBI address bank 3 */
 
-#define EBI_CS0      (1 << 1) /**< EBI chip select line 0 */
-#define EBI_CS1      (1 << 2) /**< EBI chip select line 1 */
-#define EBI_CS2      (1 << 3) /**< EBI chip select line 2 */
-#define EBI_CS3      (1 << 4) /**< EBI chip select line 3 */
+#define EBI_CS0      (uint32_t)(1 << 1) /**< EBI chip select line 0 */
+#define EBI_CS1      (uint32_t)(1 << 2) /**< EBI chip select line 1 */
+#define EBI_CS2      (uint32_t)(1 << 3) /**< EBI chip select line 2 */
+#define EBI_CS3      (uint32_t)(1 << 4) /**< EBI chip select line 3 */
 
 /*******************************************************************************
  ********************************   ENUMS   ************************************
